requirements.txt
pip-chill==1.0.3
pyinstaller==5.12.0
python-vlc==3.0.18122
